complete project

+++ app/Http/Controllers/StudentController.php
@@ -16,7 +16,7 @@ class StudentController extends Controller
 {
     public function index(Request $request)
     {
-        // Gate::authorize('student_access');
+        Gate::authorize('student_access');
         $students = Student::search($request)->paginate(10);
         $classes = ClassResource::collection(Classes::all());
         return Inertia::render("Student/Index", [
@@ -27,12 +27,14 @@ class StudentController extends Controller
         ]);
     }
     public function create(){
+        Gate::authorize('student_create');
         $classes = ClassResource::collection(Classes::all());
         return Inertia::render("Student/Create",[
             'classes' => $classes 
         ]);
     }
     public function edit(Student $student ){
+        Gate::authorize('student_edit');
         $classes = ClassResource::collection(Classes::all());
         return Inertia::render("Student/Edit",[
             'classes' => $classes,
@@ -47,16 +49,19 @@ class StudentController extends Controller
     }
     public function store(StoreStudentRequest $request)
     {
+        Gate::authorize('student_create');
         Student::create($request->all());
         return redirect(route("students.index"))->with("success","");
     }
     public function update(UpdateStudentRequest $request, Student $student)
     {
+        Gate::authorize('student_edit');
         $student->update($request->all());
         return redirect(route("students.index"))->with("success","");
     }
     public function destroy(Student $student)
     {
+        Gate::authorize('student_delete');
         $student->delete();
         return redirect(route("students.index"))->with("success","");
     }
